channels_demo/requirements.txt
Django==3.2
channels==3.0.5
channels-redis==3.4.1
uvicorn==0.18.3
gunicorn==19.10.0
djangorestframework==3.13.1